Make dropdown component more reusable
